Added automatic documentation generation
I've written a script that complements Ciaran's download script.
This will take the contents of all sample code files and match them with their respective markdown file. If one does not exist, it will create one.

Once it has opened the md file, it dumps the python code in with the required formatting to make it display properly. It also puts a file header.

There are a lot of changed files, but that's fine. I just had to delete all the markdown files to let the script make new ones as some of the names didn't quite match correctly.